Refactored router and routes files
